migrations

+++ database/migrations/2023_09_18_040113_add_columns_to_site_general_settings_table.php
@@ -13,9 +13,11 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::table('site_general_settings', function (Blueprint $table) {
-            $table->dropColumn('delivery_status');
-        });
+        if (Schema::hasColumn('site_general_settings', 'delivery_status')) {
+            Schema::table('site_general_settings', function (Blueprint $table) {
+                $table->dropColumn('delivery_status');
+            });
+        }
     }
 
     /**
